+++ resolwe/storage/connectors/localconnector.py
@@ -50,6 +50,41 @@ class LocalFilesystemConnector(BaseStorageConnector):
                 f.write(data)
                 data_remaining = len(data) == chunk_size
 
+    def multipart_push_start(self, url, size=None):
+        """Start a multipart upload.
+
+        :returns: the upload id.
+        :raises AssertionError: when url exists or upload size is not given.
+        """
+        assert size is not None, f"{self}: size of multipart upload must be specified."
+        path = self.base_path / url
+        assert not path.exists(), f"Path {url} already exists, aborting upload."
+        if not path.is_file():
+            with path.open("wb") as stream:
+                stream.truncate(size)
+        return url
+
+    def multipart_push(self, upload_id, url, part_number, chunk_size, data, md5=None):
+        """Upload a single part of multipart upload."""
+        path = self.base_path / url
+        assert path.is_file(), f"{self}: multipart upload file {path} does not exist."
+        with path.open("r+b") as stream:
+            stream.seek((part_number - 1) * chunk_size)
+            shutil.copyfileobj(data, stream, 1024 * 1024)
+        return dict()
+
+    def multipart_push_complete(self, upload_id, url, completed_chunks):
+        """Complete the multipart push."""
+        path = self.base_path / url
+        assert path.is_file(), f"{self}: multipart upload file {path} does not exist."
+        return dict()
+
+    def multipart_push_abort(self, upload_id, url):
+        """Abort multiport upload."""
+        path = self.base_path / url
+        if path.is_file():
+            path.unlink()
+
     @validate_url
     def get(self, url, stream, chunk_size=BaseStorageConnector.CHUNK_SIZE):
         """Get data from the given URL and write it into the given stream."""